# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Library Bookshelf API",
    "description": "API for books and authors information, including user authentication.",
    "version": "1.0.0"
  },
  "host": "library-bookshelf.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "sessionAuth": {
      "type": "apiKey",
      "name": "Cookie",
      "in": "header",
      "description": "Session cookie for authentication. This cookie is generated by express-session."
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Register a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User registration details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "john_doe"
                },
                "email": {
                  "type": "string",
                  "example": "john@example.com"
                },
                "password": {
                  "type": "string",
                  "example": "yourPassword"
                }
              },
              "required": [
                "username",
                "email",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Invalid input or user already exists"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Log in a user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "username": {
                  "type": "string",
                  "example": "john_doe"
                },
                "password": {
                  "type": "string",
                  "example": "yourPassword"
                }
              },
              "required": [
                "username",
                "password"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful, session cookie set"
          },
          "400": {
            "description": "Authentication failed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "description": "Log out the current user",
        "responses": {
          "200": {
            "description": "Logout successful"
          },
          "400": {
            "description": "Logout failed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "sessionAuth": []
          }
        ]
      }
    },
...
  },
...
}